Goodnews bug solved
