+++ dart prjoect/fixed/bin/fixed.dart
@@ -1,0 +1,5 @@
+import 'package:fixed/fixed.dart' as fixed;
+
+void main(List<String> arguments) {
+  print('Hello world: ${fixed.calculate()}!');
+}
