+++ dart prjoect/fixed/lib/fixed.dart
@@ -1,0 +1,3 @@
+int calculate() {
+  return 6 * 7;
+}
