+++ dart prjoect/fixed/test/fixed_test.dart
@@ -1,0 +1,8 @@
+import 'package:fixed/fixed.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('calculate', () {
+    expect(calculate(), 42);
+  });
+}
